components/reader/code/requirements.txt
opencv-python==4.4.0.44
parse==1.18.0
psycopg2-binary==2.8.6
pytest==6.1.2
textdistance==4.2.0
pyAesCrypt==0.4.3